Create short-lived kubeconfig for integration test cluster (rebase) (#1)
* fix logging (run-int-tests) (#819)

* Performance Config (#818)

* Deployer config

* Landscaper config

* Metrics in Helm Chart Templates and Values (#821)

* Metrics

* Test (run-int-tests)

* Component Library Flag (#822)

* Flag for component library

* Generate

* Pass component library flag to subobjects

* Default component library

* Generate

* Mark fields as optional (run-int-tests)

* Revert "Component Library Flag (#822)" (#824)

This reverts commit 63a77b10280842ff0d71649d540daec389bbfad4.

* add feedback to proposal (#826)

* OCM Flag (#825)

* OCM flag in context

* OCM flag in landscaper config

* Generate

* OCM flag in deployer config

* Generate (run-int-tests)

* Proposal timeouts and typical error page (#823)

* timeout proposal

* Release v0.78.0

* Prepare next Dev Cycle v0.79.0-dev

* update pipeline definition (#829)

* fix blueprint docu (#832)

* Fix pipeline (#835)

* adjust pipeline to run with specific component lib version
(run-int-tests)

* minor adjustments

* (run-int-tests)

* update pipeline definition for head updates (#830)

* update pipeline definition for head updates

* update pipeline for release jobs

* fix dependency relationship for update release

* Pipeline fix (#836)

* fix pipeline

* (run-int-tests)

* Update .ci/pipeline_definitions


---------


* true in pipeline definition converted to True (#839)

* Integrate ocm lib only  (#837)

* ocm-lib-test (run-int-tests)

* (run-int-tests)

---------


* USE_OCM_LIB parameter for local int tests (#846)

---------

+++ hack/testcluster/pkg/shootcluster_manager.go
@@ -41,6 +41,9 @@ const (
 	// prPrefix is the second part of shoot clusters used for integration tests
 	prPrefix = "pr"
 
+	ocmlibIdentifier = "o"
+	cnudieIdentifier = "c"
+
 	localStartPrefix = namePrefix + prPrefix + "0-"
 	headUpdatePrefix = namePrefix + prPrefix + "1-"
 
@@ -77,10 +80,12 @@ type ShootClusterManager struct {
 	numClustersStartDeleteOldest int
 	durationForClusterDeletion   time.Duration
 	prID                         string
+	useOCMLib                    bool
 }
 
 func NewShootClusterManager(log utils.Logger, gardenClusterKubeconfigPath, namespace,
-	authDirectoryPath string, maxNumOfClusters, numClustersStartDeleteOldest int, durationForClusterDeletion, prID string) (*ShootClusterManager, error) {
+	authDirectoryPath string, maxNumOfClusters, numClustersStartDeleteOldest int, durationForClusterDeletion, prID string,
+	useOCMLib bool) (*ShootClusterManager, error) {
 
 	log.Logfln("Create cluster manager with:")
 	log.Logfln("  GardenClusterKubeconfigPath: " + gardenClusterKubeconfigPath)
@@ -90,6 +95,7 @@ func NewShootClusterManager(log utils.Logger, gardenClusterKubeconfigPath, names
 	log.Logfln("  NumClustersStartDeleteOldest: " + strconv.Itoa(numClustersStartDeleteOldest))
 	log.Logfln("  DurationForClusterDeletion: " + durationForClusterDeletion)
 	log.Logfln("  PrID: " + prID)
+	log.Logfln("  UseOCMLib: " + strconv.FormatBool(useOCMLib))
 
 	duration, err := time.ParseDuration(durationForClusterDeletion)
 	if err != nil {
@@ -105,6 +111,7 @@ func NewShootClusterManager(log utils.Logger, gardenClusterKubeconfigPath, names
 		numClustersStartDeleteOldest: numClustersStartDeleteOldest,
 		durationForClusterDeletion:   duration,
 		prID:                         prID,
+		useOCMLib:                    useOCMLib,
 	}, nil
 }
 
@@ -262,7 +269,13 @@ func (o *ShootClusterManager) checkAndDeleteExistingTestShoots(ctx context.Conte
 }
 
 func (o *ShootClusterManager) generateShootName() string {
-	return namePrefix + prPrefix + o.prID + "-" + strconv.Itoa(rng.Intn(9000)+1000)
+	var libID string
+	if o.useOCMLib {
+		libID = ocmlibIdentifier
+	} else {
+		libID = cnudieIdentifier
+	}
+	return namePrefix + libID + "-" + prPrefix + o.prID + "-" + strconv.Itoa(rng.Intn(9000)+1000)
 }
 
 func (o *ShootClusterManager) matchesNamePattern(name string) bool {
@@ -548,10 +561,13 @@ func (o *ShootClusterManager) deleteOutdatedShootCluster(ctx context.Context, ga
 		clusterNamePrefix := clusterName[:len(clusterName)-4]
 		newClusterNamePrefix := newClusterName[:len(newClusterName)-4]
 
+		libID := clusterName[3:4]
+		newLibID := newClusterName[3:4]
+
 		if hasDeletionTimestamp {
 			continue
 		} else if !strings.HasPrefix(clusterName, localStartPrefix) && !strings.HasPrefix(clusterName, headUpdatePrefix) &&
-			clusterNamePrefix == newClusterNamePrefix {
+			clusterNamePrefix == newClusterNamePrefix && libID == newLibID {
 			o.log.Logfln("test shoot cluster %s will be deleted because a new test for the same PR is triggered", clusterName)
 
 			if err = o.deleteShootCluster(ctx, gardenClientForShoots, gardenClientForSecrets, clusterName); err != nil {
